+++ mllaunchpad/datasources.py
@@ -571,12 +571,19 @@ class FileDataSource(DataSource):
             expires: 0    # generic parameter, see documentation on DataSources
             tags: [train] # generic parameter, see documentation on DataSources and DataSinks
             options: {}   # used as **kwargs when fetching the data using `pandas.read_csv`
+            dtypes_path: ./some/file.dtypes # optional: location with the csv's column dtypes info
           my_raw_datasource:
             type: text_file  # raw files can also be of type `binary_file`
             path: /some/file.txt  # Can be URL
             expires: 0    # generic parameter, see documentation on DataSources
             tags: [train] # generic parameter, see documentation on DataSources and DataSinks
             options: {}   # used as **kwargs when fetching the data using `fh.read`
+
+    When loading `csv` or `euro_csv` type formats, you can use the setting `dtypes_path`
+    to specify a location with dtypes description for the csv (usually generated earlier
+    by using :class:`FileDataSink`'s `dtypes_path` setting). These dtypes will be enforced when
+    reading the csv, which helps avoid problems when `pandas.read_csv` interprets data differently than you do.
+    Use `dtypes_path` to enforce dtype parity between csv datasinks and datasources.
 
     Using the raw formats `binary_file` and `text_file`, you can read arbitrary data, as long as
     it can be represented as a `bytes` or a `str` object, respectively. Please note that while possible, it is not
@@ -617,6 +624,7 @@ class FileDataSource(DataSource):
 
         self.type = ds_type
         self.path = datasource_config["path"]
+        self.dtypes_path = datasource_config.get("dtypes_path")
 
     def get_dataframe(
         self, params: Dict = None, chunksize: Optional[int] = None
@@ -640,10 +648,23 @@ class FileDataSource(DataSource):
         kw_options = self.options
 
         logger.debug(
-            "Loading type {} file {} with chunksize {} and options {}...".format(
-                self.type, self.path, chunksize, kw_options
+            "Loading type {} file {} with dtypes_path {}, chunksize {} and options {}...".format(
+                self.type, self.path, self.dtypes_path, chunksize, kw_options
             )
         )
+        if self.dtypes_path is not None:
+            input_dtypes = pd.read_csv(self.dtypes_path).set_index("columns")
+            kw_options["dtype"] = {
+                item: input_dtypes.loc[item]["dtypes"]
+                for item in input_dtypes.index
+                if input_dtypes.loc[item]["dtypes"] != "datetime"
+            }
+            kw_options["parse_dates"] = [
+                item
+                for item in input_dtypes.index
+                if input_dtypes.loc[item]["dtypes"] == "datetime"
+            ]
+
         if self.type == "csv":
             df = pd.read_csv(self.path, chunksize=chunksize, **kw_options)
         elif self.type == "euro_csv":
@@ -721,11 +742,18 @@ class FileDataSink(DataSink):
             path: /some/file.csv  # Can be URL, uses `df.to_csv` internally
             tags: [train] # generic parameter, see documentation on DataSources and DataSinks
             options: {}   # used as **kwargs when fetching the data using `df.to_csv`
+            dtypes_path: ./some/file.dtypes # optional: location for saving the csv's column dtypes info
           my_raw_datasink:
             type: text_file  # raw files can also be of type `binary_file`
             path: /some/file.txt  # Can be URL
             tags: [train] # generic parameter, see documentation on DataSources and DataSinks
             options: {}   # used as **kwargs when writing the data using `fh.write`
+
+    When saving `csv` or `euro_csv` type formats, you can use the setting `dtypes_path`
+    to specify a location where to save dtypes descriptions for the csv (that you can use later
+    with :class:`FileDataSource`'s `dtypes_path` setting). These dtypes will be enforced when
+    reading the csv, which helps avoid problems when `pandas.read_csv` interprets data differently than you do.
+    Use `dtypes_path` to enforce dtype parity between csv datasinks and datasources.
 
     Using the raw formats `binary_file` and `text_file`, you can persist arbitrary data, as long as
     it can be represented as a `bytes` or a `str` object, respectively. Please note that while possible, it is not
@@ -765,6 +793,7 @@ class FileDataSink(DataSink):
 
         self.type = ds_type
         self.path = datasink_config["path"]
+        self.dtypes_path = datasink_config.get("dtypes_path")
 
     def put_dataframe(
         self,
@@ -772,7 +801,7 @@ class FileDataSink(DataSink):
         params: Dict = None,
         chunksize: Optional[int] = None,
     ) -> None:
-        """Write a pandas dataframe to file.
+        """Write a pandas dataframe to file and optionally the dtypes if included in the configuration.
         The default is not to save the dataframe's row index.
         Configure the DataSink's `options` dict to pass keyword arguments to `my_df.to_csv`.
 
@@ -797,10 +826,22 @@ class FileDataSink(DataSink):
             kw_options["index"] = False
 
         logger.debug(
-            "Writing dataframe to type {} file {} with options {}...".format(
-                self.type, self.path, kw_options
+            "Writing dataframe to type {} file {} with options {} and dtypes_path {}...".format(
+                self.type, self.path, kw_options, self.dtypes_path
             )
         )
+        if self.dtypes_path:
+            dtypes_file = pd.DataFrame(
+                dataframe.dtypes, columns=["dtypes"]
+            ).rename_axis("columns")
+            dtypes_file.loc[
+                dtypes_file["dtypes"] == "object", "dtypes"
+            ] = "str"
+            dtypes_file.loc[
+                dtypes_file["dtypes"] == "datetime64[ns]", "dtypes"
+            ] = "datetime"
+            dtypes_file.to_csv(self.dtypes_path)
+
         if self.type == "csv":
             dataframe.to_csv(self.path, **kw_options)
         elif self.type == "euro_csv":